[VarDumper] Allow HtmlDumper users to directly get the header
If the HtmlDumper is being embedded as part of a larger component, it
may not be desirable to have the HtmlDumper automatically dump the
header containing static assets.

Give HtmlDumper users more control by:

* Making getDumpHeader a public function, so that they can directly get
  the header without having to dump a variable.

* Add markHeaderAsDumped: this enables users to dump variables first
  without the header, if it's more convenient that way.  They can then
  call getDumpHeader at a later time.

Motivation: I'm integrating this with PHP Debug Bar; not having direct
access to the dump header makes it difficult to properly include these
assets in the page header along with the other debug bar static assets.
Furthermore, actual variable dumps are only stored in string variables
until they are later presented to the user via AJAX; we'd rather not
have the dump header bundled with that.