Add new feature to serializer: default_constructor_arguments
You can now add empty_data to the context on deserialization
of objects. This allow you to deserialize objects that have
requirements in the constructor that can't be given.

Basically, it helps you hydrate with value objects, so the
validation component can invalid the object without the
serializer send an error.