Adds an "AUTHENTICATION_SUCCESS_SENSITIVE" event that includes sensitive user credentials, dispatched immediately prior to the existing "AUTHENTICATION_SUCCESS" event (which does not include sensitive user credentials, by default).
The existing "AUTHENTICATION_SUCCESS" event can be configured such that the token is not sanitized of
sensitive data, but that introduces the problem of exposing this data to *all* authentication success
subscribers/listeners when the vast majority do not need, and should not have, access to such data.
The newly added "AUTHENTICATION_SUCCESS_SENSITIVE" event includes the unsanitized token, leaving the
raw password/api-key), enabling actions such as password re-hashing and other credentials-aware tasks.

Between the existing and newly added authentication events, there is now a *clear* separation between
subscribers/listeners that *should* have user credentials (and must shoulder the added responsibility
of handling it), and those that *should not* have access to user credentials.