Improve compatibility of output formatter
preg_match with offset capture returns string offsets in bytes. If
mbstring function overloading is enabled this means subsequent
calculations on characters will not line up.

In practice this causes parts of the formatting tags to appear inline in
the formatted string.

This updates the code to use mb_string function with encoding explicitly
set to ASCII. This works for servers where function overloading is
enabled and servers where it is disabled.