Invalidate Translator cache when new translation files are added
Just before writing the Translator cache, find out in which directories
the translation files were found, and add those directories as resources
for the catalogue so that when a new file appears in one of those
directories the catalogue cache will be invalidated.

This may cause some false positives in case non-translation files are in
those directories, but since the recommendation is to keep translations
on their own, it should not happen.