[Form] add more values to `false_values` of CheckboxType
When consuming a Symfony Form through an API, like manually calling the URL with form data in the URL or Request Body, `?my_checkbox=false` or `?my_checkbox=off` will not work.

We already encountered this _issue_ on 2/3 projects, because we were using `?my_checkbox=false` and it was evaluated as `true` before me configure `invalid_values` array.

To prevent this, I think it would be nice to add more false values to the CheckboxType type.

What do you think about it?

Thanks