Fix #32396
The src/Symfony/Bridge/Doctrine/Form/DoctrineOrmTypeGuesser.php implements
the FormTypeGuesserInterface and that interface is only defined in the
symfony/form component.
By implementing that interface that code depends on it not only for test but
also in production use.

We discovered it because we have an api which doesn't require symfony form
who blew up with the update to php 7.2.20.
The release of php 7.2.20 fixed a bug that was allowing the implementation
of interface that are not available.