Automatically route messages when `from_transport` tag is set
When you have multiple transports, it can become quite cumbersome to map every message to all the transports that it needs.

Instead, you can tag each handler with `from_transport` to explicitly say on which transport it should handle. Having both the transport and the message that the handler will support, we can automatically add the routing too.