Add a donation section in the README
Hello there, 

As a crypto fan, I would love to see a Bitcoin address inside the readme to support the SF project. 
I didn't know how to propose this, so I just added this line inside the README file. 
Also, of course the bitcoin address has to be created by the team (to get the private keys).
If you have any questions, let's talk about it here !