Optimization: Remove unnecessary else and elseif blocks
This is a rather big PR, but makes the same change throughout the code base.
In all changes in this PR, it removes `else` and `elseif` blocks where they are not necessary.

```php
if ($x === 1) {
	return 'Foo';
}
else {
	return 'Bar';
}
```

In the snippet above, we don't really need to use an `else` block because the block right above it stops the code flow with the `return` statement.

```diff
if ($x === 1) {
	return 'Foo';
}
-else {
-	return 'Bar';
-}
+
+ return 'Bar';
}

```

This makes the code more cleaner and more easier to read.

Throughout this PR, it simplifies `else` and `elseif` blocks if the previous block stops the code flow with `return`, `throw`, `continue`, `break`, or `yield` statements.

This optimization is more pronounced when elseif chains are removed.

```php

if ($value instanceof Parameter) {
    return $this->getParameterCall((string) $value);
} elseif ($value instanceof Expression) {
    return $this->getExpressionCall((string) $value);
} elseif ($value instanceof Definition) {
    return new TaggedValue('service', (new Parser())->parse("_:\n".$this->addService('_', $value), Yaml::PARSE_CUSTOM_TAGS)['_']['_']);
} elseif ($value instanceof AbstractArgument) {
    return new TaggedValue('abstract', $value->getText());
} elseif (\is_object($value) || \is_resource($value)) {
    throw new RuntimeException('Unable to dump a service container if a parameter is an object or a resource.');
}
```

This block, without a lot of breathing room, and seemingly connected chain is not exactly "connected" because all blocks halt the flow with their `return` calls.

```php
if ($value instanceof Reference) {
    return $this->getServiceCall((string) $value, $value);
}

if ($value instanceof Parameter) {
    return $this->getParameterCall((string) $value);
}

if ($value instanceof Expression) {
    return $this->getExpressionCall((string) $value);
}

if ($value instanceof Definition) {
    return new TaggedValue('service', (new Parser())->parse("_:\n".$this->addService('_', $value), Yaml::PARSE_CUSTOM_TAGS)['_']['_']);
}

if ($value instanceof AbstractArgument) {
    return new TaggedValue('abstract', $value->getText());
}

if (\is_object($value) || \is_resource($value)) {
    throw new RuntimeException('Unable to dump a service container if a parameter is an object or a resource.');
}
```

With more breathing room and clear separation, it is more easier for eyes to scan through individual cases.