[Serializer] Add a @Mapping Annotation
Add a @Mapping annotation applicable to a class to configure the groups, the maxDepth, and the serializedName of a set of attributes. This could be used to ease readability when a class own a lot of attributes and the same groups are applied on a recurring basis. This also ease to configure attributes embedded in the traits used by this class, since you can apply some class-specific groups to them (currently a little bit tricky id you don't use the yaml / xml format).

The "ignore" metadata is currently not supported, simply for readability reasons (an @IgnoreAttributes annotation would probably be better) bet could easily implemented.

The "maxDepth" parameter on the root of the annotation could also be removed, to avoid unnecessary metadata storage for attributes not using it.