[Mailer] send cloned Message to the bus
This more or less reverts 829566cdea90a0be0231a, but it restores the
behaviour in 4.4: The **cloned** Mime Message is dispatched to the
messenger instead of the original one.

This allows us to change the message using an event listener and then
get that *changed* instance in the message handler.