Encode slashes and adapt UrlGenerator closer to RFC 3986
This is a proof of concept for https://github.com/symfony/symfony/issues/13017 and addresses some pressing issues when comparing the current implementation with RFC 3986.

If this would be adapted in Symfony, the changed behaviour probably should become optional as some changes will break backwards compatibility.
For now this should be fine to point out the changed behaviour.