[Validator] Add optional conditionExpression to Count Assertion.
With the conditionExpression option for the count assertion, it is possible to count only those elements in the iterable, that match the given symfony expression language. Each item from the iterable will be passed as variable 'item' into the expression language.

Add CountValidatorCollectionTest to test the new feature against an iterable type that is not an array.
Let test classes for array and collection extend the new CountValidatorIterable test to only run the new tests for iterable types.