Don't limit retries of toolbar loading
On projects where finishing the request takes some time, loading
should not be stopped after a definitive amount of time.

The loading bar will only be shown after the original amount of
requests (5) were unsuccessful to flash as little as possible.

Also the hold-off has been changed to 500 and a maximum of 2000.