[ConsoleComponent] Implemented support for output sections larger than terminal
The current console section utility implementation was not aware of the output terminals dimensions causing a variety of rendering glitches when trying to overwrite section contents that are larger then the displaying terminal.

Root of the problem is the inability to move the cursor above the terminal home position using ANSI/VT100 control escape sequences. The implementation has been extended to compute and apply partial differences limited to the visible part of the terminal.