BUG: revert negatable ansi
When I create a command and configure it with the addOption 'no-ansi', the current 5.3 is throwing an exception.

The reason behind it is, that it is configure as negatable. When it is adding it check if the no-ansi is exists. Then it throws an exception.
In the 5.2 situation it then checked if the options equals. If not throw an exception.

https://github.com/symfony/console/blob/5.3/Input/InputDefinition.php#L255 part where the check is coming for the negatable and throws an exception.
https://github.com/symfony/console/blob/5.3/Input/InputDefinition.php#L233 The equals check where in the past the error was prevented.

Reverting the option to two values, seems like the right solution.