Fix HttpClientInterface::withOptions declaration
This method is not static and thus should not be declared as static.
The return type is dynamic though. I don't know how it could be declared
as a dynamic return type but it looks enough to me as the caller knows
the type of object it manipulates. Using this methodd should not break
the original typing detected by the IDE or static analysis tools.

See https://docs.phpdoc.org/3.0/guide/references/phpdoc/tags/method.html#method