feat: fix asset package availability detection
If the package assets is available in the dependencies, Symfony will
enable automatically assets. Otherwise assets are disabled. The problem
is that sometimes you want to explicetely enable assets features (in the
case of testing an application you may have assets as **dev**
dependency).

In the case assets are enable explicitely, the service assets.packages
is defined and we should not remove the definition
twig.extension.assets.