[Messenger] Allow fine-grained control in log severity of failures
This commit adds a new interface method on the `RetryStrategyInterface`
allowing custom retry strategies to control the log severity of failed
and retried messages.

This is a spin-off from https://github.com/symfony/symfony/pull/43160

| Q             | A
| ------------- | ---
| Branch?       | 6.0
| Bug fix?      | no
| BC break? | yes - custom retry strategies require including the new trait or implementing the new method
| New feature?  | yes - changelog not yet updated
| Deprecations? | no
| Tickets       | -
| License       | MIT
| Doc PR        | not yet, collecting feedback first

The default logic for determining the log severity for failed messages
has been (since https://github.com/symfony/symfony/pull/43492):

 - recoverable exceptions are logged as warning
 - unrecoverable exceptions are logged as critical
 - neutral exceptions are logged as warning after a non-final attempt
 - neutral exceptions are logged as critical after a final attempt

This PR implements the exact same behaviour but in a way that allows
users to augment or override the default behaviour with their own
specific needs. For example, having specific exceptions logged as
`EMERGENCY` or to log them as `NOTICE` until or including the last
attempt.

The refactoring done to make this possible is:

 - move the logic determining if an exception is
   recoverable/unrecoverable/neutral to the `HandlerFailedException`
   so it can be used by retry strategy objects

 - add a new interface method on the `RetryStrategyInterface` (BC
   break!)

 - implement the existing behaviour using the new mechanism in a
   trait, this way, custom retry strategies can easily re-use the
   default behaviour