Add `exclude` to TaggedIterator and TaggedLocator
With this change, you can exclude one or more services from a tagged iterator
or locator.

It's common when working with Delegator or Chain services that call a list of services that
implement an interface, to also want to implement the interface on Delegator/Chain.

An example of this is in the Symfony HttpKernel component:

https://github.com/symfony/symfony/blob/0d6e859db236e37b8baa2b2bf4c8b7d14d151570/src/Symfony/Component/HttpKernel/CacheClearer/CacheClearerInterface.php#L19-L25
https://github.com/symfony/symfony/blob/0d6e859db236e37b8baa2b2bf4c8b7d14d151570/src/Symfony/Component/HttpKernel/CacheClearer/ChainCacheClearer.php#L21-L31

The `ChainCacheClearer` implements `CacheClearerInterface` and calls a list of `CacheClearerInterface`
clearers.

If that would have been configured with `#[AutoconfigureTag]` and `#[TaggedIterator]`, the
`ChainCacheClearer` would receive itself in the `iterable $clearers`.

With this change, it can exclude itself and rely fully on autowire/configure.

Another example:

```php
#[AutoconfigureTag]
interface ErrorTracker
{
    public function trackError(string $error): void;
}

final class SentryErrorTracker implement ErrorTracker
{
    public function trackError(string $error): void
    {
        echo "Send error to Sentry\n";
    }
}

final class NewRelicErrorTracker implement ErrorTracker
{
    public function trackError(string $error): void
    {
        echo "Send error to NewRelic\n";
    }
}

final class DelegatingErrorTracker implements ErrorTracker
{
    public function __construct(
        #[TaggedIterator(ErrorTracker::class, exclude: self::class)]
        private iterable $trackers
    ) {}

    public function trackError(string $error): void
    {
        foreach($this->trackers as $tracker) {
            $tracker->trackError($error);
        }
    }
}

// Alias ErrorTracker interface to DelegatingErrorTracker service

final class MyController
{
    public function __construct(private ErrorTracker $errorTracker) {}
    public function __invoke()
    {
        $this->errorTracker->trackError('Hello, World!');
    }
}
```

Without this change, the `DelegatingErrorTracker` would receive itself in the tagged
iterator.