Fixes max_duration not being respected
As reported on #46316, when a request gets a successful response in further attempts, the max_duration is not respected and the request may take longer than the value specified here.

My way to solve the issue was by reducing the timeout option in future requests, making it smaller in every attempt. By doing so, the total time of a $client->request will never be much longer than max_duration. Perhaps a second or two maybe, due to the waiting time between requests, but not as much as the timeout.