[Security/Http] Throw `AccessDeniedException` in `UserValueResolver` when user is required but not authenticated
When you add `#[CurrentUser] UserInterface $user` to a controller and the user is
not authenticated, the following error is produced:

```
Cannot autowire argument $user of "App\Controller::myAction()": it references class "App\User" but no such service exists.
```

This error doesn't help with resolving the problem.

The `UserValueResolver` should check and see if the argument is non-optional and immediately
stop execution by throwing an `AccessDeniedException`. This can help redirect the user to a
login page.