Update TraceableAccessDecisionManager.php
Removed the "if" statement that verifies if the ADM is instance of AccessDecisionManager.

This control breaks profiler when it is trying to browse "Security" tab while a custom ADM was developed. The reason is that the "if" statement prevents SecurityDataCollector.php to retrieve the strategy and the list of voters.

Based on my tests this modification doesn't affect the regular functionality of security-core.