Introduce `#[AsAlias]` attribute
Sometimes we want to alias something in the container with another service. This is usually done in less visible config files like `services.php` or a `Bundle` file. To make these aliases configurations more visible, the `#[AsAlias]` attribute is introduced. The attribute can be used on class that you want your e.g. Interface to alias to.

This allows doing:

```php
#[AsAlias(alias: 'bar')]
class HelloWorldService
{
}
```

instead of

```
$services->setAlias(‘bar’, HelloWorldService::class);
```