Updated MonologBridge\Logger so it doesn't extend Monolog\Logger
This shouldn't break any BC. An adapter for the public Monolog\Logger methods has been created as a trait. The trait has been created as "deprecated", so it can be removed later.

This change depends on a change in the upstream Monolog class. An MR has been opened for monolog to cover that