Extract the EmojiTransliterator to a dedicated package
Due to the size of the data needed to transliterate emojis, using a
dedicated package allows to include the package in the project only when
needed.
In Symfony 6.x, the symfony/intl component depends on this component for
BC reasons. Root projects can implement a replace hack in their
composer.json to skip this package if they don't need it. This
dependency will be removed in Symfony 7.0.