Introduce HandlerDescriptorInterface
The `HandlerDescriptor` class is final and therefore cannot be extended. To make this a bit more
flexible we introduce the `HandlerDescriptorInterface` that is used everywhere.

It also introduces a `.messenger.handler_descriptor_class` parameter that's used in the
MessengerPass. With this parameter it's possible to use a custom implementation of the
`HandlersLocator`.

This is an alternative approach to https://github.com/symfony/symfony/pull/50980.