fix ReflectionMethod with php8.3 in MicroKernelTrait
the issue is this https://github.com/php/php-src/issues/9470
When used with php 8.3RC1 KernelTestcases will not work do to the method
configureContainer not beeing visible to the $this. i am a little
confused, if this is a bug in php or should be adressed here. the fix
is simple and straight forward.