Configure mockclient if mock_response_factory has been set on a scoped client.
This makes it possible to use different mock response factories on different services and also to easily test services that uses http client endpoints.

It also generates an extra test service http_client.mock_client.$name that can be used to configure the MockHttpClient instance if you want to inject a different responsefactory.

We also change the framework.http_client.mock_client_factory attribute to allow for a boolean
so you can just mock without injecting a responsefactory in the service setup.